refactor(ai): isolate logic into separate files
