+++ packages/ai-commands/src/sql/chat.ts
@@ -1,0 +1,78 @@
+import { SchemaBuilder } from '@serafin/schema-builder'
+import { OpenAIStream } from 'ai'
+import { codeBlock, oneLine, stripIndent } from 'common-tags'
+import { jsonrepair } from 'jsonrepair'
+import OpenAI from 'openai'
+import { ContextLengthError, EmptyResponseError, EmptySqlError } from '../errors'
+import { Message } from '../types'
+
+/**
+ * Responds to a conversation about writing a SQL query.
+ *
+ * @returns A `ReadableStream` containing the response text and SQL.
+ */
+export async function chatSql(
+  openai: OpenAI,
+  messages: Message[],
+  existingSql?: string,
+  entityDefinitions?: string[]
+): Promise<ReadableStream<Uint8Array>> {
+  const initMessages: OpenAI.Chat.Completions.ChatCompletionMessageParam[] = [
+    {
+      role: 'system',
+      content: stripIndent`
+      The generated SQL (must be valid SQL).
+      - For primary keys, always use "id bigint primary key generated always as identity" (not serial)
+      - Prefer creating foreign key references in the create statement
+      - Prefer 'text' over 'varchar'
+      - Prefer 'timestamp with time zone' over 'date'
+      - Use vector(384) data type for any embedding/vector related query
+      - Always use double apostrophe in SQL strings (eg. 'Night''s watch')
+      - Always use semicolons
+      - Output as markdown
+      - Always include code snippets if available
+      `,
+    },
+  ]
+
+  if (entityDefinitions) {
+    const definitions = codeBlock`${entityDefinitions.join('\n\n')}`
+    initMessages.push({
+      role: 'user',
+      content: oneLine`Here is my database schema for reference: ${definitions}`,
+    })
+  }
+
+  if (existingSql !== undefined && existingSql.length > 0) {
+    const sqlBlock = codeBlock`${existingSql}`
+    initMessages.push({
+      role: 'user',
+      content: codeBlock`
+        Here is the existing SQL I wrote for reference:
+        ${sqlBlock}
+      `.trim(),
+    })
+  }
+
+  if (messages) {
+    initMessages.push(...messages)
+  }
+
+  try {
+    const response = await openai.chat.completions.create({
+      model: 'gpt-3.5-turbo-0125',
+      messages: initMessages,
+      max_tokens: 1024,
+      temperature: 0,
+      stream: true,
+    })
+
+    // Transform the streamed SSE response from OpenAI to a ReadableStream
+    return OpenAIStream(response)
+  } catch (error) {
+    if (error instanceof Error && 'code' in error && error.code === 'context_length_exceeded') {
+      throw new ContextLengthError()
+    }
+    throw error
+  }
+}
