+++ packages/ai-commands/src/sql/functions.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, test } from '@jest/globals'
 import { codeBlock } from 'common-tags'
 import OpenAI from 'openai'
-import { formatSql } from '../test/util'
-import { debugSql, titleSql } from './sql'
+import { formatSql } from '../../test/util'
+import { debugSql, titleSql } from './functions'
 
 const openAiKey = process.env.OPENAI_API_KEY
 const openai = new OpenAI({ apiKey: openAiKey })
