+++ packages/ai-commands/src/sql/index.ts
@@ -1,0 +1,2 @@
+export * from './chat'
+export * from './functions'
