+++ packages/ai-commands/src/sql/rls.test.ts
@@ -15,8 +15,8 @@ import {
   renderJsonExpression,
   renderTargets,
   unwrapNode,
-} from '../test/sql-util'
-import { collectStream, extractMarkdownSql, withMetadata } from '../test/util'
+} from '../../test/sql-util'
+import { collectStream, extractMarkdownSql, withMetadata } from '../../test/util'
 import { chatRlsPolicy } from './rls'
 
 const openAiKey = process.env.OPENAI_API_KEY
